+++ src/main/java/com/anthonyhilyard/merchantmarkers/MerchantMarkersConfig.java
@@ -32,6 +32,7 @@ public class MerchantMarkersConfig
 	public static final ForgeConfigSpec SPEC;
 	public static final MerchantMarkersConfig INSTANCE;
 
+	public final BooleanValue alwaysShow;
 	public final BooleanValue showThroughWalls;
 	public final BooleanValue showArrow;
 	public final BooleanValue showOnMiniMap;
@@ -108,6 +109,7 @@ public class MerchantMarkersConfig
 	{
 		build.comment("Client Configuration").push("client").push("visual_options");
 
+		alwaysShow = build.comment(" If markers above villagers should always show. If false, they will only show when the configured keybind is held.").define("always_show", true);
 		showThroughWalls = build.comment(" If markers should be visible through walls and other obstructions.").define("show_through_walls", true);
 		showArrow = build.comment(" If markers should include an arrow under the profession-specific icon.").define("show_arrow", true);
 		showOnMiniMap = build.comment(" If icons should show on minimaps. (Currently supports Xaero's Minimap, FTB Chunks, and JourneyMap).").define("show_on_minimap", true);
@@ -182,7 +184,6 @@ public class MerchantMarkersConfig
 	{
 		if (event.getConfig().getModId().equals(Loader.MODID))
 		{
-			Loader.LOGGER.info("Merchant Markers config reloaded.");
 			Markers.clearResourceCache();
 			try
 			{