+++ src/main/java/com/anthonyhilyard/merchantmarkers/compat/JourneymapHandler.java
@@ -13,7 +13,6 @@ import journeymap.client.api.event.ClientEvent;
 import journeymap.client.api.event.forge.EntityRadarUpdateEvent;
 import net.minecraft.client.Minecraft;
 import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.merchant.villager.AbstractVillagerEntity;
 import net.minecraft.util.text.StringTextComponent;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -51,7 +50,7 @@ public class JourneymapHandler implements IClientPlugin
 			LivingEntity entity = event.getWrappedEntity().getEntityLivingRef().get();
 
 			// If this entity is marker-able, update the texture before drawing.
-			if (entity instanceof AbstractVillagerEntity && !(entity).isBaby())
+			if (Markers.shouldShowMarker(entity))
 			{
 				String profession = Markers.getProfessionName(entity);