+++ src/main/java/com/anthonyhilyard/merchantmarkers/mixin/FTBChunksMapIconWidgetMixin.java
@@ -13,12 +13,15 @@ import dev.ftb.mods.ftbchunks.client.MapIconWidget;
 import dev.ftb.mods.ftbchunks.client.MapType;
 import dev.ftb.mods.ftbchunks.integration.MapIcon;
 import net.minecraft.entity.Entity;
+import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraft.util.ResourceLocation;
 
-
+@SuppressWarnings("null")
 @Mixin(MapIconWidget.class)
 public class FTBChunksMapIconWidgetMixin
 {
+	private final static ResourceLocation VILLAGER_LOCATION = new ResourceLocation("villager");
+
 	@Redirect(method = "draw", at = @At(value = "INVOKE", target = "Ldev/ftb/mods/ftbchunks/integration/MapIcon;draw(Ldev/ftb/mods/ftbchunks/client/MapType;Lcom/mojang/blaze3d/matrix/MatrixStack;IIIIZ)V", remap = false), remap = false)
 	public void redirectIconDraw(MapIcon icon, MapType mapType, MatrixStack stack, int x, int y, int w, int h, boolean outsideVisibleArea)
 	{
@@ -28,7 +31,7 @@ public class FTBChunksMapIconWidgetMixin
 			Entity entity = entityIcon.entity;
 
 			// If this is a villager, return the dynamic texture instead of the default one.
-			if (MerchantMarkersConfig.INSTANCE.showOnMiniMap.get() && entity.getType().getRegistryName().equals(new ResourceLocation("villager")))
+			if (MerchantMarkersConfig.INSTANCE.showOnMiniMap.get() && ForgeRegistries.ENTITIES.getKey(entity.getType()).equals(VILLAGER_LOCATION))
 			{
 				FTBChunksHandler.setCurrentEntity(entity);
 			}